maybe the last version, extracted view
